reorganized player and gameplay scripts
