--- Assets/_Game/Scripts/Level.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1268c662bdf89cd449604a24c8cb97e1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
