+++ Assets/_Game/Scripts/Play/Damageable.cs
@@ -1,9 +1,7 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Events;
 
-namespace Game
+namespace Game.Play
 {
     public class Damageable : MonoBehaviour
     {
