+++ Assets/_Game/Scripts/Play/Items/HealingItem.cs
@@ -1,8 +1,9 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using Game.Player;
 
-namespace Game.Items
+namespace Game.Play.Items
 {
     public class HealingItem : ItemBase
     {
