+++ Assets/_Game/Scripts/Play/Items/SpeedUpItem.cs
@@ -1,13 +1,12 @@
 using System.Collections;
 using System.Collections.Generic;
-using Game.Items.Statistics;
+using Game.Play.Items.Statistics;
 using UnityEngine;
 using Game.Player;
-using Game.Weapons;
 
-namespace Game.Items
+namespace Game.Play.Items
 {
-    public class DamageUpItem : ItemBase
+    public class SpeedUpItem : ItemBase
     {
         [SerializeField]
         StatModifier _modifier;
@@ -15,7 +14,7 @@ namespace Game.Items
         protected override void Pickup(PlayerStats player)
         {
             base.Pickup(player);
-            player.Damage.AddModifier(_modifier);
+            player.Speed.AddModifier(_modifier);
         }
     }
 }
