+++ Assets/_Game/Scripts/Play/Level/Room.cs
@@ -2,14 +2,14 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using Game.Camera;
+using Game.Play.Camera;
 using UnityEngine;
 
 #if UNITY_EDITOR
 using UnityEditor;
 #endif
 
-namespace Game.Level
+namespace Game.Play.Level
 {
     public class Room : MonoBehaviour
     {
