+++ Assets/_Game/Scripts/Player/PlayerWeapon.cs.meta
