+++ Assets/_Game/Scripts/UI/EndMenu.cs
@@ -1,6 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
-using Game.Level;
+using Game.Play.Level;
 using Game.Utility;
 using UnityEngine;
 using UnityEngine.SceneManagement;
