+++ Assets/_Game/Scripts/UI/HUDHealth.cs
@@ -1,4 +1,6 @@
 using System;
+using Game.Play.Level;
+using Game.Play;
 using UnityEngine;
 using UnityEngine.UI;
