+++ Assets/_Game/Scripts/UI/StartMenu.cs
@@ -1,6 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
-using Game.Level;
+using Game.Play.Level;
 using UnityEngine;
 using UnityEngine.SceneManagement;
