+++ Assets/_Game/Scripts/Weapons/WeaponBomb.cs
@@ -1,6 +1,8 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using Game.Play.Level;
+using Game.Play;
 using UnityEngine;
 
 namespace Game.Weapons
