+++ Assets/_Game/Scripts/Weapons/WeaponCollision.cs
@@ -1,5 +1,7 @@
 using System.Collections;
 using System.Collections.Generic;
+using Game.Play.Level;
+using Game.Play;
 using UnityEngine;
 
 namespace Game.Weapons
